+++ ext/rbkit_event.h
@@ -22,19 +22,19 @@ typedef struct _rbkit_event_header {
 
 typedef struct _rbkit_obj_created_event {
   rbkit_event_header event_header;
-  unsigned long object_id;
+  unsigned long long object_id;
   const char *klass;
   rbkit_allocation_info *allocation_info;
 } rbkit_obj_created_event;
 
-rbkit_obj_created_event *new_rbkit_obj_created_event(unsigned long object_id, const char *klass, rbkit_allocation_info *info);
+rbkit_obj_created_event *new_rbkit_obj_created_event(unsigned long long object_id, const char *klass, rbkit_allocation_info *info);
 
 typedef struct _rbkit_obj_destroyed_event {
   rbkit_event_header event_header;
-  unsigned long object_id;
+  unsigned long long object_id;
 } rbkit_obj_destroyed_event;
 
-rbkit_obj_destroyed_event *new_rbkit_obj_destroyed_event(unsigned long object_id);
+rbkit_obj_destroyed_event *new_rbkit_obj_destroyed_event(unsigned long long object_id);
 
 typedef struct _rbkit_hash_event {
   rbkit_event_header event_header;
